feat(etl): redesign workers by filing importance into 3 tiers
Reorganize 20 workers into 39 tiered workers to prioritize valuable
filings first:
- Tier 1 (01-17): SEC Primary (10-K/10-Q + stock prices), one year each
- Tier 2 (18-22): Non-SEC schemas (econ, census, geo, crime, weather)
- Tier 3 (23-39): SEC Secondary (8-K, DEF 14A, insider), one year each

Add generate_sec_primary_model, generate_sec_secondary_model, and
generate_single_schema_model functions to common.sh.

Also fix NumberFormatException in XbrlToParquetConverter where XBRL
order attributes contain decimal values like "1.0" that fail
Integer.parseInt — use Double.parseDouble cast to int instead.

+++ govdata/src/main/java/org/apache/calcite/adapter/govdata/sec/XbrlToParquetConverter.java
@@ -2450,7 +2450,7 @@ public class XbrlToParquetConverter implements FileConverter {
       String order = arc.getAttribute("order");
       if (order != null && !order.isEmpty()) {
         try {
-          data.put("order", Integer.parseInt(order));
+          data.put("order", (int) Double.parseDouble(order));
         } catch (NumberFormatException e) {
           data.put("order", null);
         }
@@ -2614,7 +2614,7 @@ public class XbrlToParquetConverter implements FileConverter {
               data.put("from_concept", cleanConceptName(fromRef));
               data.put("to_concept", cleanConceptName(toRef));
               data.put("weight", weight != null && !weight.isEmpty() ? Double.parseDouble(weight) : null);
-              data.put("order", order != null && !order.isEmpty() ? Integer.parseInt(order) : i);
+              data.put("order", order != null && !order.isEmpty() ? (int) Double.parseDouble(order) : i);
               data.put("preferred_label", ixRel.getAttribute("preferredLabel"));
               dataList.add(data);
               processedRelationships.add(relationshipKey);
@@ -2973,7 +2973,7 @@ public class XbrlToParquetConverter implements FileConverter {
           String order = element.getAttribute("order");
           if (order != null && !order.isEmpty()) {
             try {
-              data.put("order", Integer.parseInt(order));
+              data.put("order", (int) Double.parseDouble(order));
             } catch (NumberFormatException e) {
               data.put("order", null);
             }
